Refactor SQLiteGraphDatabase -> SQLDatabase and SQLiteGraphDatabase
